+++ src/main/java/com/monito/domains/container/dto/request/ContainerMetricsRequestDTO.java
@@ -44,10 +44,10 @@ public class ContainerMetricsRequestDTO {
     private Long txBytes;
     private Long rxPackets;
     private Long txPackets;
-    private Long rxErrors;
-    private Long txErrors;
-    private Long rxDropped;
-    private Long txDropped;
+    private Integer rxErrors;
+    private Integer txErrors;
+    private Integer rxDropped;
+    private Integer txDropped;
 
     // Storage 관련
     private Long sizeRw;         // Read-Write Layer 크기
